code/src/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["GHuiKai <aodesai19899278@13.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pymongo = "^4.6.1"
boto3 = "^1.33.11"
streamlit = "^1.29.0"
azure-storage-blob = "^12.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"